docs: update build instructions in README
The current build instructions are outdated. For example, they ask
the user to "Switch to the CombinedLargeEngine branch", which cannot
be found in the repository. As well,
8bdd205d03bfadae901be8e7d1d396f867f940f3 adds dependency management,
which should be mentioned in the build instructions.

Update the existing build instructions to remove outdated steps, add
new steps, and improve language.